Support up to 2^16 colors & select the newly imported set by default

+++ src/builder/graphics/VoxelWorld.ts
@@ -76,7 +76,7 @@ export default class VoxelWorld {
         let cell = this.cells[cellId];
         if (!cell) {
             const {cellSize} = this;
-            cell = new Uint8Array(cellSize * cellSize * cellSize);
+            cell = new Uint16Array(cellSize * cellSize * cellSize);
             this.cells[cellId] = cell;
         }
         return cell;
